chore: refactor admin-api handling into ee folder
