+++ web/src/features/entitlements/constants/entitlements.ts
@@ -19,6 +19,7 @@ const entitlements = [
   "data-retention",
   "prompt-protected-labels",
   "custom-dashboards",
+  "admin-api",
 ] as const;
 export type Entitlement = (typeof entitlements)[number];
 
@@ -97,6 +98,7 @@ export const entitlementAccess: Record<
       "cloud-multi-tenant-sso",
       "integration-blobstorage",
       "prompt-protected-labels",
+      "admin-api",
     ],
     entitlementLimits: {
       "annotation-queue-count": false,
@@ -115,6 +117,7 @@ export const entitlementAccess: Record<
       "cloud-multi-tenant-sso",
       "integration-blobstorage",
       "prompt-protected-labels",
+      "admin-api",
     ],
     entitlementLimits: {
       "annotation-queue-count": false,
@@ -167,6 +170,7 @@ export const entitlementAccess: Record<
       "audit-logs",
       "data-retention",
       "prompt-protected-labels",
+      "admin-api",
     ],
     entitlementLimits: {
       "annotation-queue-count": false,
