+++ web/src/pages/api/admin/organizations/[organizationId]/apiKeys/index.ts
@@ -1,21 +1,14 @@
 import { type NextApiRequest, type NextApiResponse } from "next";
-import { prisma } from "@langfuse/shared/src/db";
 import { logger } from "@langfuse/shared/src/server";
 import { AdminApiAuthService } from "@/src/features/admin-api/server/adminApiAuth";
-import { auditLog } from "@/src/features/audit-logs/auditLog";
-import { z } from "zod";
-import { createAndAddApiKeysToDb } from "@langfuse/shared/src/server/auth/apiKeys";
-
-const validateQueryAndExtractId = (query: unknown): string | null => {
-  const inputQuerySchema = z.object({
-    organizationId: z.string(),
-  });
-  const validation = inputQuerySchema.safeParse(query);
-  if (!validation.success) {
-    return null;
-  }
-  return validation.data.organizationId;
-};
+import {
+  validateQueryAndExtractId,
+  handleGetApiKeys,
+  handleCreateApiKey,
+} from "@/src/ee/features/admin-api/organizations/apiKeys";
+import { prisma } from "@langfuse/shared/src/db";
+import { hasEntitlementBasedOnPlan } from "@/src/features/entitlements/server/hasEntitlement";
+import { getSelfHostedInstancePlanServerSide } from "@/src/features/entitlements/server/getPlan";
 
 export default async function handler(
   req: NextApiRequest,
@@ -30,6 +23,17 @@ export default async function handler(
     // Verify admin API authentication, but allow non-langfuse cloud use-cases
     if (!AdminApiAuthService.handleAdminAuth(req, res, false)) {
       return;
+    }
+
+    if (
+      !hasEntitlementBasedOnPlan({
+        plan: getSelfHostedInstancePlanServerSide(),
+        entitlement: "admin-api",
+      })
+    ) {
+      return res.status(403).json({
+        error: "This feature is not available on your current plan.",
+      });
     }
 
     const organizationId = validateQueryAndExtractId(req.query);
@@ -49,9 +53,9 @@ export default async function handler(
     // Handle different HTTP methods
     switch (req.method) {
       case "GET":
-        return await handleGet(req, res, organizationId);
+        return await handleGetApiKeys(req, res, organizationId);
       case "POST":
-        return await handlePost(req, res, organizationId);
+        return await handleCreateApiKey(req, res, organizationId);
       default:
         res.status(405).json({ error: "Method Not Allowed" });
         return;
@@ -60,77 +64,4 @@ export default async function handler(
     logger.error("Failed to process organization API key request", e);
     res.status(500).json({ error: "Internal server error" });
   }
-}
-
-async function handleGet(
-  req: NextApiRequest,
-  res: NextApiResponse,
-  organizationId: string,
-) {
-  const apiKeys = await prisma.apiKey.findMany({
-    where: {
-      orgId: organizationId,
-      scope: "ORGANIZATION",
-    },
-    select: {
-      id: true,
-      createdAt: true,
-      expiresAt: true,
-      lastUsedAt: true,
-      note: true,
-      publicKey: true,
-      displaySecretKey: true,
-    },
-    orderBy: {
-      createdAt: "asc",
-    },
-  });
-
-  return res.status(200).json({ apiKeys });
-}
-
-async function handlePost(
-  req: NextApiRequest,
-  res: NextApiResponse,
-  organizationId: string,
-) {
-  // Validate the request body
-  const createApiKeySchema = z.object({
-    note: z.string().optional(),
-  });
-
-  const validationResult = createApiKeySchema.safeParse(req.body);
-
-  if (!validationResult.success) {
-    return res.status(400).json({
-      error: "Invalid request body",
-      details: validationResult.error.format(),
-    });
-  }
-
-  const { note } = validationResult.data;
-
-  // Create the API key
-  const apiKeyMeta = await createAndAddApiKeysToDb({
-    prisma,
-    entityId: organizationId,
-    note,
-    scope: "ORGANIZATION",
-  });
-
-  // Log the API key creation
-  await auditLog({
-    resourceType: "apiKey",
-    resourceId: apiKeyMeta.id,
-    action: "create",
-    orgId: organizationId,
-    orgRole: "ADMIN",
-    apiKeyId: "ADMIN_KEY",
-  });
-
-  logger.info(
-    `Created API key ${apiKeyMeta.id} for organization ${organizationId} via admin API`,
-  );
-
-  return res.status(201).json(apiKeyMeta);
 }
