+++ web/src/pages/api/public/organizations/memberships/index.ts
@@ -2,16 +2,13 @@ import { ApiAuthService } from "@/src/features/public-api/server/apiAuth";
 import { cors, runMiddleware } from "@/src/features/public-api/server/cors";
 import { prisma } from "@langfuse/shared/src/db";
 import { logger, redis } from "@langfuse/shared/src/server";
-import { Role } from "@langfuse/shared";
-import { z } from "zod";
+import {
+  handleGetMemberships,
+  handleUpdateMembership,
+} from "@/src/ee/features/admin-api/public/organizations/memberships";
 
 import { type NextApiRequest, type NextApiResponse } from "next";
-
-// Schema for request body validation
-const MembershipSchema = z.object({
-  userId: z.string(),
-  role: z.nativeEnum(Role),
-});
+import { hasEntitlementBasedOnPlan } from "@/src/features/entitlements/server/hasEntitlement";
 
 export default async function handler(
   req: NextApiRequest,
@@ -51,13 +48,24 @@ export default async function handler(
     });
   }
 
+  if (
+    !hasEntitlementBasedOnPlan({
+      plan: authCheck.scope.plan,
+      entitlement: "admin-api",
+    })
+  ) {
+    return res.status(403).json({
+      error: "This feature is not available on your current plan.",
+    });
+  }
+
   // Route to the appropriate handler based on HTTP method
   try {
     switch (req.method) {
       case "GET":
-        return handleGet(req, res, authCheck.scope.orgId);
+        return handleGetMemberships(req, res, authCheck.scope.orgId);
       case "PUT":
-        return handlePut(req, res, authCheck.scope.orgId);
+        return handleUpdateMembership(req, res, authCheck.scope.orgId);
       default:
         // This should never happen due to the check at the beginning
         return res.status(405).json({
@@ -73,88 +81,4 @@ export default async function handler(
       error: "Internal server error",
     });
   }
-}
-
-// GET - Retrieve all organization memberships
-async function handleGet(
-  req: NextApiRequest,
-  res: NextApiResponse,
-  orgId: string,
-) {
-  const memberships = await prisma.organizationMembership.findMany({
-    where: {
-      orgId: orgId,
-    },
-    include: {
-      user: {
-        select: {
-          id: true,
-          email: true,
-          name: true,
-        },
-      },
-    },
-  });
-
-  return res.status(200).json({
-    memberships: memberships.map((membership) => ({
-      userId: membership.userId,
-      role: membership.role,
-      email: membership.user.email,
-      name: membership.user.name,
-    })),
-  });
-}
-
-// PUT - Update or create an organization membership
-async function handlePut(
-  req: NextApiRequest,
-  res: NextApiResponse,
-  orgId: string,
-) {
-  const validatedBody = MembershipSchema.safeParse(req.body);
-  if (!validatedBody.success) {
-    return res.status(400).json({
-      error: "Invalid request body",
-      details: validatedBody.error.errors,
-    });
-  }
-
-  // Check if user exists
-  const user = await prisma.user.findUnique({
-    where: {
-      id: validatedBody.data.userId,
-    },
-  });
-
-  if (!user) {
-    return res.status(404).json({
-      error: "User not found",
-    });
-  }
-
-  // Upsert the membership
-  const membership = await prisma.organizationMembership.upsert({
-    where: {
-      orgId_userId: {
-        orgId,
-        userId: validatedBody.data.userId,
-      },
-    },
-    update: {
-      role: validatedBody.data.role,
-    },
-    create: {
-      orgId,
-      userId: validatedBody.data.userId,
-      role: validatedBody.data.role,
-    },
-  });
-
-  return res.status(200).json({
-    userId: membership.userId,
-    role: membership.role,
-    email: user.email,
-    name: user.name,
-  });
 }
