+++ web/src/pages/api/public/projects/[projectId]/apiKeys/[apiKeyId].ts
@@ -2,23 +2,12 @@ import { type NextApiRequest, type NextApiResponse } from "next";
 import { prisma } from "@langfuse/shared/src/db";
 import { logger, redis } from "@langfuse/shared/src/server";
 import { ApiAuthService } from "@/src/features/public-api/server/apiAuth";
-import { auditLog } from "@/src/features/audit-logs/auditLog";
-import { z } from "zod";
 import { cors, runMiddleware } from "@/src/features/public-api/server/cors";
-
-const validateQueryParams = (
-  query: unknown,
-): { projectId: string; apiKeyId: string } | null => {
-  const inputQuerySchema = z.object({
-    projectId: z.string(),
-    apiKeyId: z.string(),
-  });
-  const validation = inputQuerySchema.safeParse(query);
-  if (!validation.success) {
-    return null;
-  }
-  return validation.data;
-};
+import {
+  validateQueryParams,
+  handleDeleteApiKey
+} from "@/src/ee/features/admin-api/public/projects/projectById/apiKeys/apiKeyById";
+import { hasEntitlementBasedOnPlan } from "@/src/features/entitlements/server/hasEntitlement";
 
 export default async function handler(
   req: NextApiRequest,
@@ -55,6 +44,17 @@ export default async function handler(
     }
     // END CHECK AUTH
 
+    if (
+      !hasEntitlementBasedOnPlan({
+        plan: authCheck.scope.plan,
+        entitlement: "admin-api",
+      })
+    ) {
+      return res.status(403).json({
+        error: "This feature is not available on your current plan.",
+      });
+    }
+
     const params = validateQueryParams(req.query);
     if (!params) {
       return res.status(400).json({ message: "Invalid request parameters" });
@@ -79,7 +79,7 @@ export default async function handler(
     // Handle different HTTP methods
     switch (req.method) {
       case "DELETE":
-        return await handleDelete(
+        return await handleDeleteApiKey(
           req,
           res,
           projectId,
@@ -94,53 +94,4 @@ export default async function handler(
     logger.error("Failed to process project API key request", e);
     res.status(500).json({ message: "Internal server error" });
   }
-}
-
-async function handleDelete(
-  req: NextApiRequest,
-  res: NextApiResponse,
-  projectId: string,
-  apiKeyId: string,
-  orgId: string,
-) {
-  // Check if API key exists and belongs to the project
-  const apiKey = await prisma.apiKey.findFirst({
-    where: {
-      id: apiKeyId,
-      projectId,
-      scope: "PROJECT",
-    },
-  });
-
-  if (!apiKey) {
-    return res.status(404).json({ message: "API key not found" });
-  }
-
-  // Delete the API key
-  const deleted = await new ApiAuthService(prisma, redis).deleteApiKey(
-    apiKeyId,
-    projectId,
-    "PROJECT",
-  );
-
-  if (!deleted) {
-    return res.status(500).json({ message: "Failed to delete API key" });
-  }
-
-  // Log the API key deletion
-  await auditLog({
-    resourceType: "apiKey",
-    resourceId: apiKeyId,
-    action: "delete",
-    orgId: orgId,
-    projectId: projectId,
-    orgRole: "ADMIN",
-    apiKeyId: "ORG_KEY",
-  });
-
-  logger.info(
-    `Deleted API key ${apiKeyId} for project ${projectId} via public API`,
-  );
-
-  return res.status(200).json({ success: true });
 }
