+++ web/src/pages/api/public/projects/[projectId]/index.ts
@@ -1,18 +1,13 @@
 import { ApiAuthService } from "@/src/features/public-api/server/apiAuth";
 import { cors, runMiddleware } from "@/src/features/public-api/server/cors";
 import { prisma } from "@langfuse/shared/src/db";
+import { logger, redis } from "@langfuse/shared/src/server";
 import {
-  logger,
-  redis,
-  QueueJobs,
-  ProjectDeleteQueue,
-  type ApiAccessScope,
-} from "@langfuse/shared/src/server";
-import { randomUUID } from "crypto";
-import { projectRetentionSchema } from "@/src/features/auth/lib/projectRetentionSchema";
+  handleUpdateProject,
+  handleDeleteProject,
+} from "@/src/ee/features/admin-api/public/projects/projectById";
 import { hasEntitlementBasedOnPlan } from "@/src/features/entitlements/server/hasEntitlement";
 import { type NextApiRequest, type NextApiResponse } from "next";
-import { projectNameSchema } from "@/src/features/auth/lib/projectNameSchema";
 
 export default async function handler(
   req: NextApiRequest,
@@ -56,6 +51,17 @@ export default async function handler(
   }
   // END CHECK AUTH
 
+  if (
+    !hasEntitlementBasedOnPlan({
+      plan: authCheck.scope.plan,
+      entitlement: "admin-api",
+    })
+  ) {
+    return res.status(403).json({
+      error: "This feature is not available on your current plan.",
+    });
+  }
+
   // Check if project exists and belongs to the organization
   const project = await prisma.project.findFirst({
     where: {
@@ -72,141 +78,10 @@ export default async function handler(
 
   // Route to the appropriate handler based on HTTP method
   if (req.method === "PUT") {
-    return handlePut(req, res, projectId, authCheck.scope);
+    return handleUpdateProject(req, res, projectId, authCheck.scope);
   }
 
   if (req.method === "DELETE") {
-    return handleDelete(req, res, projectId, authCheck.scope);
-  }
-}
-
-async function handlePut(
-  req: NextApiRequest,
-  res: NextApiResponse,
-  projectId: string,
-  scope: ApiAccessScope,
-) {
-  try {
-    const { name, retention } = req.body;
-
-    // Validate project name
-    try {
-      projectNameSchema.parse({ name });
-    } catch (error) {
-      return res.status(400).json({
-        message: "Invalid project name. Should be between 3 and 60 characters.",
-      });
-    }
-
-    // Validate retention days using the schema
-    try {
-      projectRetentionSchema.parse({ retention });
-    } catch (error) {
-      return res.status(400).json({
-        message: "Invalid retention value. Must be 0 or at least 7 days.",
-      });
-    }
-
-    // If retention is non-zero, check for data-retention entitlement
-    if (retention > 0) {
-      const hasDataRetentionEntitlement = hasEntitlementBasedOnPlan({
-        entitlement: "data-retention",
-        plan: scope.plan,
-      });
-
-      if (!hasDataRetentionEntitlement) {
-        return res.status(403).json({
-          message:
-            "The data-retention entitlement is required to set a non-zero retention period.",
-        });
-      }
-    }
-
-    // Update the project with the new retention setting
-    const updatedProject = await prisma.project.update({
-      where: {
-        id: projectId,
-        orgId: scope.orgId,
-      },
-      data: {
-        name,
-        retentionDays: retention,
-      },
-      select: {
-        id: true,
-        name: true,
-        retentionDays: true,
-      },
-    });
-
-    return res.status(200).json({
-      id: updatedProject.id,
-      name: updatedProject.name,
-      ...(updatedProject.retentionDays // Do not add if null or 0
-        ? { retentionDays: updatedProject.retentionDays }
-        : {}),
-    });
-  } catch (error) {
-    logger.error("Failed to update project", error);
-    return res.status(500).json({ message: "Internal server error" });
-  }
-}
-
-async function handleDelete(
-  req: NextApiRequest,
-  res: NextApiResponse,
-  projectId: string,
-  scope: ApiAccessScope,
-) {
-  try {
-    // API keys need to be deleted from cache. Otherwise, they will still be valid.
-    await new ApiAuthService(prisma, redis).invalidateProjectApiKeys(projectId);
-
-    // Delete API keys from DB
-    await prisma.apiKey.deleteMany({
-      where: {
-        projectId: projectId,
-        scope: "PROJECT",
-      },
-    });
-
-    // Mark project as deleted
-    await prisma.project.update({
-      where: {
-        id: projectId,
-        orgId: scope.orgId,
-      },
-      data: {
-        deletedAt: new Date(),
-      },
-    });
-
-    // Queue project deletion job
-    const projectDeleteQueue = ProjectDeleteQueue.getInstance();
-    if (!projectDeleteQueue) {
-      logger.error("ProjectDeleteQueue is not available");
-      return res.status(500).json({
-        message: "Internal server error",
-      });
-    }
-
-    await projectDeleteQueue.add(QueueJobs.ProjectDelete, {
-      timestamp: new Date(),
-      id: randomUUID(),
-      payload: {
-        projectId: projectId,
-        orgId: scope.orgId,
-      },
-      name: QueueJobs.ProjectDelete,
-    });
-
-    return res.status(202).json({
-      success: true,
-      message:
-        "Project deletion has been initiated and is being processed asynchronously",
-    });
-  } catch (error) {
-    logger.error("Failed to delete project", error);
-    return res.status(500).json({ message: "Internal server error" });
+    return handleDeleteProject(req, res, projectId, authCheck.scope);
   }
 }
